requirements.txt
Pillow
opencv-python
tkinter